🧰: fix styling of dragged inspector property
